GRAPHICS: Split BlendBlitImpl_Default into a separate file
